minecraft/pyproject.toml
[project]
name = "minecraft"
version = "0.1.0"
description = "minecraft"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot2>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-adapter-telegram>=0.1.0b20",
    "nonebot-adapter-minecraft>=1.8.0",
    "nonebot-adapter-discord>=1.1.3",
    "nonebot-plugin-mcqq>=2.9.0",
]

[tool.nonebot]
plugin_dirs = ["plugins"]
builtin_plugins = ["echo", "single_session"]
[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-mcqq = ["nonebot_plugin_mcqq"]


[tool.nonebot.adapters]
"@local" = []
nonebot-adapter-onebot = [{name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11"}]
nonebot-adapter-telegram = [{name = "Telegram", module_name = "nonebot.adapters.telegram"}]
nonebot-adapter-minecraft = [{name = "Minecraft", module_name = "nonebot.adapters.minecraft"}]
nonebot-adapter-discord = [{name = "Discord", module_name = "nonebot.adapters.discord"}]
